# Conflicts:
#	src/octoprint/filemanager/storage/local.py
#	src/octoprint/logging/handlers.py
#	src/octoprint/plugin/__init__.py
#	src/octoprint/plugins/virtual_printer/virtual.py
#	src/octoprint/printer/standard.py
#	src/octoprint/server/__init__.py
#	src/octoprint/server/api/connection.py
#	src/octoprint/server/api/files.py
#	src/octoprint/settings.py
#	src/octoprint/static/js/app/viewmodels/connection.js
#	src/octoprint/util/__init__.py
#	tests/plugin/test_settings.py

+++ src/octoprint/server/api/timelapse.py
@@ -1,5 +1,5 @@
-# coding=utf-8
-from __future__ import absolute_import, division, print_function
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 __author__ = "Gina Häußge <osd@foosel.net>"
 __license__ = 'GNU Affero General Public License http://www.gnu.org/licenses/agpl.html'
@@ -16,7 +16,7 @@ import octoprint.util as util
 from octoprint.settings import settings, valid_boolean_trues
 
 from octoprint.server import admin_permission, printer
-from octoprint.server.util.flask import redirect_to_tornado, require_firstrun, get_json_command_from_request, with_revalidation_checking
+from octoprint.server.util.flask import redirect_to_tornado, no_firstrun_access, get_json_command_from_request, with_revalidation_checking
 from octoprint.server.api import api
 
 from octoprint.server import NO_CONTENT
@@ -66,9 +66,12 @@ def _etag(unrendered, lm=None):
 
 	import hashlib
 	hash = hashlib.sha1()
-	hash.update(str(lm))
-	hash.update(repr(config))
-	hash.update(repr(_DATA_FORMAT_VERSION))
+	def hash_update(value):
+		value = value.encode('utf-8')
+		hash.update(value)
+	hash_update(str(lm))
+	hash_update(repr(config))
+	hash_update(repr(_DATA_FORMAT_VERSION))
 
 	return hash.hexdigest()
 
@@ -76,7 +79,7 @@ def _etag(unrendered, lm=None):
 @with_revalidation_checking(etag_factory=lambda lm=None: _etag(request.values.get("unrendered", "false") in valid_boolean_trues, lm=lm),
                             lastmodified_factory=lambda: _lastmodified(request.values.get("unrendered", "false") in valid_boolean_trues),
                             unless=lambda: request.values.get("force", "false") in valid_boolean_trues)
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_LIST.require(403)
 def getTimelapseData():
 	timelapse = octoprint.timelapse.current
@@ -113,6 +116,7 @@ def getTimelapseData():
 		finished_list.append(output)
 
 	result = dict(config=config,
+	              enabled=settings().getBoolean(["webcam", "timelapseEnabled"]),
 	              files=finished_list)
 
 	if unrendered:
@@ -122,31 +126,33 @@ def getTimelapseData():
 
 
 @api.route("/timelapse/<filename>", methods=["GET"])
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_DOWNLOAD.require(403)
 def downloadTimelapse(filename):
 	return redirect_to_tornado(request, url_for("index") + "downloads/timelapse/" + filename)
 
 
 @api.route("/timelapse/<filename>", methods=["DELETE"])
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_DELETE.require(403)
 def deleteTimelapse(filename):
-	if util.is_allowed_file(filename, ["mpg", "mpeg", "mp4", "m4v", "mkv"]):
-		timelapse_folder = settings().getBaseFolder("timelapse")
-		full_path = os.path.realpath(os.path.join(timelapse_folder, filename))
-		if full_path.startswith(timelapse_folder) and os.path.exists(full_path):
-			try:
-				os.remove(full_path)
-			except Exception as ex:
-				logging.getLogger(__file__).exception("Error deleting timelapse file {}".format(full_path))
-				return make_response("Unexpected error: {}".format(ex), 500)
+	timelapse_folder = settings().getBaseFolder("timelapse")
+	full_path = os.path.realpath(os.path.join(timelapse_folder, filename))
+	if octoprint.timelapse.valid_timelapse(full_path) \
+			and full_path.startswith(timelapse_folder) \
+			and os.path.exists(full_path) \
+			and not util.is_hidden_path(full_path):
+		try:
+			os.remove(full_path)
+		except Exception as ex:
+			logging.getLogger(__file__).exception("Error deleting timelapse file {}".format(full_path))
+			return make_response("Unexpected error: {}".format(ex), 500)
 
 	return getTimelapseData()
 
 
 @api.route("/timelapse/unrendered/<name>", methods=["DELETE"])
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_DELETE.require(403)
 def deleteUnrenderedTimelapse(name):
 	octoprint.timelapse.delete_unrendered_timelapse(name)
@@ -154,7 +160,7 @@ def deleteUnrenderedTimelapse(name):
 
 
 @api.route("/timelapse/unrendered/<name>", methods=["POST"])
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_ADMIN.require(403)
 def processUnrenderedTimelapseCommand(name):
 	# valid file commands, dict mapping command name to mandatory parameters
@@ -175,7 +181,7 @@ def processUnrenderedTimelapseCommand(name):
 
 
 @api.route("/timelapse", methods=["POST"])
-@require_firstrun
+@no_firstrun_access
 @Permissions.TIMELAPSE_ADMIN.require(403)
 def setTimelapseConfig():
 	data = request.get_json(silent=True)
